% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bivariate_estimation.R
\name{biLocal}
\alias{biLocal}
\title{Estimate the local correlation between two variables when the marginals are standard normal.}
\usage{
biLocal(data, grid, h)
}
\arguments{
\item{data}{The data matrix, one row per observation. The marginals are assumed to be standard normal.
They can be transformed using the transLocal-function.}

\item{grid}{The grid where the local correlation will be estimated.}

\item{h}{The vector of bandwidths. Length must be 2.}
}
\value{
A list with 4 elements: $data is the data, $grid is the grid, $par.est is a vector of the
estimated local correlations, and $f.est are the local likelihood density estimates there.
}
\description{
Estimate the local correlation between two variables when the marginals are standard normal.
}
\examples{
data <- cbind(rt(100, df = 10), rt(100, df = 10))
grid <- matrix(c(-1, -1, 0, 0, 1, 1), ncol = 2, byrow = TRUE)
trans <- transLocal(data = data, grid = grid)

estimates <- biLocal(data = trans$transformed.data,
                     grid = trans$transformed.grid,
                     h = c(1, 1))
}

