% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multivariate_estimation.R
\name{condLocal}
\alias{condLocal}
\title{Estimate conditional densities using the LGDE}
\usage{
condLocal(data, cond, bandwidths = NULL, gsize = 15,
  grid = apply(as.matrix(data[, 1:(dim(data)[2] - length(cond))]), 2,
  function(x) seq(quantile(x, 0.001), quantile(x, 0.999), length.out = gsize)))
}
\arguments{
\item{data}{The data matrix, one row per observation.}

\item{cond}{A vector of values to be conditioned on.}

\item{bandwidths}{A matrix of bandwidths. Must be in the same format as is
produced by the HLocal-function.}

\item{gsize}{If grid is not provided, the density will be estimated at
gsize points going diagonally through the data.}

\item{grid}{The grid at which to estimate the density. Must be a matrix with
the same number of columns as the data matrix.}
}
\value{
A list with 1 element:
    $f.est.cond is the estimated conditional density.
}
\description{
Estimate conditional densities using the LGDE
}
\examples{
# Three dimensional example. The conditioning variables must be the last columns in the data matrix.
Estimating the density of X1|X2 = 0, X3 = 0:
data <- cbind(rt(100, df = 10), rt(100, df = 10), rt(100, df = 10))
condLocal(data, cond = c(0,0))
}

