% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bandwidth_selection.R
\name{HLocal}
\alias{HLocal}
\title{Calculate the global cross-validated bandwidths for the LGDE.}
\usage{
HLocal(data)
}
\arguments{
\item{data}{The data matrix, one row per observation.}
}
\value{
A matrix. The first two columns contain each variable pair,
the last two columns contain the two bandwidths for the pair
in question.
}
\description{
Calculate the global cross-validated bandwidths for the LGDE.
}
\examples{
data <- cbind(rnorm(100), rnorm(100))
HLocal(data)
}

