% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bandwidth_selection.R
\name{kLocal}
\alias{kLocal}
\title{Uses cross validation to determine the k for kNN bandwidth selection.}
\usage{
kLocal(data, test)
}
\arguments{
\item{data}{The data matrix.}

\item{test}{Vector of integers that shall be tested. Chooses the k that has the highest likelihood.}
}
\value{
Matrix with 4 columns, one for each pair of variables. The first two columns indicate the
pair in question, the third column givs the selected k, and the last column returns a 1 if the selected
k is an endpoint for the "test"-vector. Perhaps one should change the search area?
}
\description{
Uses cross validation to determine the k for kNN bandwidth selection.
}
\examples{
data <- cbind(rnorm(100), rnorm(100))
kLocal(data, test = seq(20, 80, 3))
}

