% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bandwidth_selection.R
\name{knnBi}
\alias{knnBi}
\title{Calculate the distance to the k'th nearest neighbour for a given bivariate data set, on a given grid.}
\usage{
knnBi(data, grid, k)
}
\arguments{
\item{data}{The data set, a nx2-matrix.}

\item{grid}{The grid, a dx2-matrix.}

\item{k}{The number of neighbours.}
}
\value{
A vector of length d containing the distance to the k'th nearest neighbour for each grid point.
}
\description{
Calculate the distance to the k'th nearest neighbour for a given bivariate data set, on a given grid.
}
\examples{
data <- cbind(rnorm(100), rnorm(100))
grid <- cbind(c(-1, 0, 1), c(-1, 0, 1))
k <- 10
knnBi(data, grid, k)
}

