\name{lgde-package}
\alias{lgde-package}
\alias{lgde}
\docType{package}
\title{
\packageTitle{lgde}
}
\description{
\packageDescription{lgde}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lgde}
\packageIndices{lgde}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{lgde}

Maintainer: \packageMaintainer{lgde}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
