% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multivariate_estimation.R
\name{multiLocal.knn}
\alias{multiLocal.knn}
\title{Estimate a multivariate density function using the LGDE and kNN bandwidths.}
\usage{
multiLocal.knn(data, k.mat = NULL, gsize = 15, grid = apply(data, 2,
  function(x) seq(quantile(x, 0.001), quantile(x, 0.999), length.out = gsize)),
  test = seq(20, 100, by = 3))
}
\arguments{
\item{data}{The data matrix, one row per observation.}

\item{k.mat}{A matrix of k. Must be in the same format as is
produced by the kLocal-function.}

\item{gsize}{If grid is not provided, the density will be estimated at
gsize points going diagonally through the data.}

\item{grid}{The grid at which to estimate the density. Must be a matrix with
the same number of columns as the data matrix.}
}
\value{
A list with 6 elements:
    $f.est are the density estimates at the grid points,
    $d is the dimension,
    $loc.cor are the local correlations,
    $k.mat is the matrix of k's that was used,
    $pairs is a two-column matrix indicating all pairs of variables,
    $grid is the grid at which the density has been estimated.
}
\description{
Estimate a multivariate density function using the LGDE and kNN bandwidths.
}
\examples{
data <- cbind(rt(100, df = 10), rt(100, df = 10))
multiLocal(data)
}

