% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bandwidth_selection.R
\name{pluginLocal}
\alias{pluginLocal}
\title{Create a bandwidth object using the plugin method h = 1.75n^{-1/6}}
\usage{
pluginLocal(n, nvar)
}
\arguments{
\item{n}{The number of observations.}

\item{nvar}{The number of variables}
}
\value{
A matrix with bandwidths to be used in the multiLocal- or condLocal-functions.
}
\description{
Create a bandwidth object using the plugin method h = 1.75n^{-1/6}
}
\examples{
data <- cbind(rnorm(100), rnorm(100))
pluginLocal(data)
}

